modified bootstrapping
