+++ index.js
@@ -1,7 +1,10 @@
 var config = require('./config');
 var role = config.ROLE || config.FOREMAN_WORKER_NAME.split(".")[0];
 console.log(role);
-//TODO parseManager and scanner should buildSets before starting
-//parseManager requires at least one parser in order to operate, it will retry until it finds at least one
-//scanner can operate, but we might leak matches if we dont have a tracked player set while scanning
-require("./"+role+".js");
+//web requires constants
+//worker requires constants (fullhistory)
+//parseManager requires constants (processparse)
+//parseManager should require parsers, it will retry until it finds at least one
+//scanner should require buildSets, we might leak matches if we dont have a tracked player set while scanning
+//retriever, parser, proxy are independent
+require("./" + role + ".js");
