+++ package.json
@@ -9,12 +9,13 @@
     "unparsed": "nf run node runner.js unparsed",
     "updateNames": "nf run node runner.js updateNames",
     "buildSets": "nf run node runner.js buildSets",
+    "constants": "nf run node runner.js constants",
     "fullhistory": "nf run node runner.js fullhistory",
     "shorthistory": "nf run node runner.js shorthistory",
     "getmissing": "nf run node runner.js getmissing",
     "webpack": "webpack -p",
     "watch": "webpack --watch",
-    "build": "npm install && npm run webpack && npm run maven",
+    "build": "npm install && npm run webpack && npm run maven && npm run constants",
     "maven": "mvn -q -f parser/pom.xml clean install -U",
     "dev": "nf -j Procfile.dev start",
     "yasp-parser": "sudo nf export -j Procfile.prod -o /etc/init/ parser && sudo restart foreman",
